Move HadoopConfigurationLocalStackUtil to parquet module
