+++ java/parquet/src/test/java/sleeper/utils/HadoopConfigurationLocalStackUtils.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.ingest.testutils;
+package sleeper.utils;
 
 import org.apache.hadoop.conf.Configuration;
 import org.testcontainers.containers.localstack.LocalStackContainer;
 
-public class HadoopConfigurationLocalStackUtil {
+public class HadoopConfigurationLocalStackUtils {
 
-    private HadoopConfigurationLocalStackUtil() {
+    private HadoopConfigurationLocalStackUtils() {
     }
 
     /**
@@ -34,7 +34,7 @@ public class HadoopConfigurationLocalStackUtil {
      */
     public static Configuration getHadoopConfiguration(LocalStackContainer container) {
         Configuration configuration = new Configuration();
-        configuration.setClassLoader(HadoopConfigurationLocalStackUtil.class.getClassLoader());
+        configuration.setClassLoader(HadoopConfigurationLocalStackUtils.class.getClassLoader());
         configuration.set("fs.s3a.endpoint", container.getEndpointOverride(LocalStackContainer.Service.S3).toString());
         configuration.set("fs.s3a.aws.credentials.provider", "org.apache.hadoop.fs.s3a.SimpleAWSCredentialsProvider");
         configuration.set("fs.s3a.access.key", container.getAccessKey());
